requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
python-multipart
sentence-transformers
faiss-cpu
numpy
pickle-mixin